Correct typo in file name
